+++ crates/convert/src/lib.rs
@@ -1,2 +1,4 @@
-pub mod poscar;
+//! Convert external file formats into Velvet data structures.
+
+mod poscar;
 pub use poscar::load_poscar;