+++ crates/convert/src/poscar.rs
@@ -1,6 +1,5 @@
-use std::convert::AsRef;
+use std::io::BufRead;
 use std::iter::FromIterator;
-use std::path::Path;
 use std::str::FromStr;
 
 use nalgebra::{Matrix3, Vector3};
@@ -10,12 +9,39 @@ use velvet_core::system::cell::Cell;
 use velvet_core::system::elements::Element;
 use velvet_core::system::{System, SystemBuilder};
 
-pub fn load_poscar<P>(path: P) -> System
+/// Returns a [`System`](velvet_core::system::System) object initialized from POSCAR formatted data.
+///
+/// # Arguments
+///
+/// * `reader` - File object or buffer to read from.
+///
+/// # Examples
+///
+/// ```
+/// use velvet_core::system::System;
+/// use velvet_convert::load_poscar;
+///
+/// let system = load_poscar("\
+///     Cubic BN
+///     3.57
+///     0.0 0.5 0.5
+///     0.5 0.0 0.5
+///     0.5 0.5 0.0
+///     B N
+///     1 1
+///     Direct
+///     0.00 0.00 0.00
+///     0.25 0.25 0.25
+///     ".as_bytes());
+///
+/// assert_eq!(system.size(), 2);
+/// ```
+pub fn load_poscar<R>(reader: R) -> System
 where
-    P: AsRef<Path>,
+    R: BufRead,
 {
     // load the poscar file
-    let poscar = Poscar::from_path(path).unwrap();
+    let poscar = Poscar::from_reader(reader).unwrap();
 
     // system size alias
     let size = poscar.num_sites();
@@ -77,11 +103,15 @@ where
 #[cfg(test)]
 mod tests {
     use super::load_poscar;
+    use std::fs::File;
+    use std::io::BufReader;
     use velvet_core::utils::test_path;
 
     #[test]
     fn argon() {
-        let sys = load_poscar(test_path("argon.poscar"));
+        let file = File::open(test_path("argon.poscar")).unwrap();
+        let reader = BufReader::new(file);
+        let sys = load_poscar(reader);
         assert_eq!(sys.size(), 108);
     }
 }